[HER-16] Follow-up Rename styled-row to index
